Move CFG/IOp tests to their project
